Abstract out common Sidekiq tracer functionality into base class
